Add files via upload
The first protocol that can be used for data transmission.

Features:

* File sharing(Uploading)
* Raw data sending
* Command sending

Pros:

File and Command sending is protected with password. That means, when you host the GCP server, it will request a password from you. Clients will be required to provide the correct password for file and command sending. Raw data can be sent without passwords.

Cons:

Traffic encryption isn't added to the protocol yet. This version, will cause problems if someone MiTMs(Man in The Middle) the network. But MiTM might be avoided because there is no broadcasting in this protocol, only sending.